feat: 주거비 dto 수정
